Refactor and Added FillRate typedef and resolver
