+++ src/graphql/typeDefs/index.ts
@@ -1,10 +1,15 @@
-import { mergeTypes } from "merge-graphql-schemas";
+import { mergeTypes } from 'merge-graphql-schemas'
 
-import report from "./report";
+//import report from "./report";
+import report from './Commons'
+import shipment from './ShipmentReport'
+import customer from './CustomerReport'
+import vendor from './VendorReport'
+import fillrate from './FillRate'
 
-const typeDefs = [report];
+const typeDefs = [report, shipment, customer, vendor, fillrate]
 
 // NOTE: 2nd param is optional, and defaults to false
 // Only use if you have defined the same type multiple times in
 // different files and wish to attempt merging them together.
-export default mergeTypes(typeDefs, { all: true });
+export default mergeTypes(typeDefs, { all: true })
